Cleanup settings components and fix password reset flow
