--- src/screens/Settings/TabGeneralSettings/DeviceSection/styles.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components/native'
-
-export const DeviceContainer = styled.View`
-  gap: 15px;
-`
