--- src/screens/Settings/TabGeneralSettings/index.jsx
@@ -1,137 +1,0 @@
-import { useState } from 'react'
-
-import { i18n } from '@lingui/core'
-import { useLingui } from '@lingui/react/macro'
-import { useNavigation } from '@react-navigation/native'
-import {
-  sendGoogleFormFeedback,
-  sendSlackFeedback
-} from 'pear-apps-lib-feedback'
-import { BackIcon } from 'pearpass-lib-ui-react-native-components'
-import { Platform, ScrollView, Text, View } from 'react-native'
-import { SafeAreaView } from 'react-native-safe-area-context'
-import Toast from 'react-native-toast-message'
-
-import { DeviceSection } from './DeviceSection'
-import { LanguageSection } from './LanguageSection'
-import { MasterPasswordSection } from './MasterPasswordSection'
-import { ReportSection } from './ReportSection'
-import { Version } from './styles'
-import { version } from '../../../../package.json'
-import { CardSingleSetting } from '../../../components/CardSingleSetting'
-import {
-  GOOGLE_FORM_KEY,
-  GOOGLE_FORM_MAPPING,
-  SLACK_WEBHOOK_URL_PATH
-} from '../../../constants/feedback'
-import { useLanguageOptions } from '../../../hooks/useLanguageOptions'
-import { ButtonLittle } from '../../../libComponents'
-import { logger } from '../../../utils/logger'
-import { settingsStyles } from '../styles'
-
-export const TabGeneralSettings = () => {
-  const { t } = useLingui()
-  const navigation = useNavigation()
-  const [language, setLanguage] = useState(i18n.locale)
-  const handleChangeLanguage = (newLang) => {
-    setLanguage(newLang)
-    i18n.activate(newLang)
-  }
-
-  const [message, setMessage] = useState('')
-  const [isLoading, setIsLoading] = useState(false)
-
-  const { languageOptions } = useLanguageOptions()
-
-  const handleReportProblem = async () => {
-    if (!message?.length) {
-      return
-    }
-
-    try {
-      setIsLoading(true)
-
-      const payload = {
-        message,
-        topic: 'BUG_REPORT',
-        app: 'MOBILE',
-        operatingSystem: Platform.OS,
-        deviceModel: Platform.constants.Brand,
-        appVersion: version
-      }
-
-      await sendSlackFeedback({
-        webhookUrPath: SLACK_WEBHOOK_URL_PATH,
-        ...payload
-      })
-
-      await sendGoogleFormFeedback({
-        formKey: GOOGLE_FORM_KEY,
-        mapping: GOOGLE_FORM_MAPPING,
-        ...payload
-      })
-
-      setMessage('')
-
-      setIsLoading(false)
-
-      Toast.show({
-        type: 'baseToast',
-        text1: t`Feedback sent`,
-        position: 'bottom',
-        bottomOffset: 100
-      })
-    } catch (error) {
-      logger.error('Error sending feedback:', error)
-
-      setIsLoading(false)
-
-      Toast.show({
-        type: 'baseToast',
-        text1: t`ERROR: Feedback not sent`,
-        position: 'bottom',
-        bottomOffset: 100
-      })
-    }
-  }
-
-  return (
-    <SafeAreaView
-      style={settingsStyles.container}
-      edges={['top', 'left', 'right']}
-    >
-      <View style={settingsStyles.header}>
-        <ButtonLittle
-          startIcon={BackIcon}
-          variant="secondary"
-          borderRadius="md"
-          onPress={() => navigation.goBack()}
-        />
-        <Text style={settingsStyles.screenTitle}>{t`General`}</Text>
-      </View>
-      <ScrollView contentContainerStyle={settingsStyles.contentContainer}>
-        <LanguageSection
-          language={language}
-          setLanguage={handleChangeLanguage}
-          title={t`Language`}
-          languageOptions={languageOptions}
-        />
-
-        <MasterPasswordSection />
-
-        <ReportSection
-          message={message}
-          setMessage={setMessage}
-          isLoading={isLoading}
-          handleReportProblem={handleReportProblem}
-        />
-
-        <DeviceSection />
-
-        <CardSingleSetting title={t`Version`}>
-          <Version>{version}</Version>
-        </CardSingleSetting>
-      </ScrollView>
-    </SafeAreaView>
-  )
-}
